requirements.txt
python-dotenv
mcp[cli]
colorama
openai
requests
streamlit
oracledb
loguru
sqlite-vec